# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShopXperience API",
    "description": "Multi-tenant Point of Sale System API",
    "version": "1.0.0",
    "contact": {
      "name": "ShopXperience Support",
      "email": "support@shopxperience.com"
    }
  },
  "servers": [
    {
      "url": "https://api.shopxperience.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "customer_id": { "type": "integer" },
          "location_id": { "type": "integer" },
          "total_amount": { "type": "number", "format": "decimal" },
          "tax_amount": { "type": "number", "format": "decimal" },
          "discount_amount": { "type": "number", "format": "decimal" },
          "payment_method": { "type": "string" },
          "status": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
...
    "/api/payments/create-payment-intent": {
      "post": {
        "summary": "Create payment intent",
        "description": "Create Stripe payment intent for card payments",
        "tags": ["Payments"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": { "type": "number", "format": "decimal" },
                  "currency": { "type": "string", "default": "usd" },
                  "payment_method_types": { "type": "array", "items": { "type": "string" }, "default": ["card"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment intent created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "client_secret": { "type": "string" },
                    "payment_intent_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequestError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/payments/confirm-payment": {
      "post": {
        "summary": "Confirm payment and create transaction",
        "description": "Confirm Stripe payment and create transaction record",
        "tags": ["Payments"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["payment_intent_id", "location_id", "items"],
                "properties": {
                  "payment_intent_id": { "type": "string" },
                  "customer_id": { "type": "integer" },
                  "location_id": { "type": "integer" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["product_id", "quantity"],
                      "properties": {
                        "product_id": { "type": "integer" },
                        "quantity": { "type": "integer" }
                      }
                    }
                  },
                  "discount_amount": { "type": "number", "format": "decimal", "default": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment confirmed and transaction created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "transaction": { "$ref": "#/components/schemas/Transaction" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequestError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequestError": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  }
}